Allow Tensor& in the unboxing logic
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42712

Previously, operators taking Tensor& as arguments or returning it couldn't be c10-full because the unboxing logic didn't support it.
This adds temporary support for that. We're planning to remove this again later, but for now we need it to make those ops c10-full.
See https://docs.google.com/document/d/19thMVO10yMZA_dQRoB7H9nTPw_ldLjUADGjpvDmH0TQ for the full plan.

This PR also makes some ops c10-full that now can be.
ghstack-source-id: 109693706

Differential Revision: [D22989242](https://our.internmc.facebook.com/intern/diff/D22989242/)